wip: refactor, add a draft general data product route
