+++ src/lib/api_app/src/api/routes/testbed/productizers/job/mod.rs
@@ -2,7 +2,7 @@ use http::{Method, StatusCode};
 use math::round;
 use std::{cmp::Ordering, collections::hash_map::DefaultHasher, env, hash::Hasher};
 
-use super::parse_testbed_request_headers;
+use crate::api::routes::testbed::testbed_utils::parse_testbed_request_headers;
 
 use app::{
     requests::engage_many_json_requests,
